@@ -64,7 +64,15 @@
                     </execution>
                 </executions>
                 <configuration>
-                    <jvmTarget>1.8</jvmTarget>
+                    <jvmTarget>17</jvmTarget>
+                </configuration>
+            </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <configuration>
+                    <source>17</source>
+                    <target>17</target>
                 </configuration>
             </plugin>
         </plugins>
@@ -72,16 +80,16 @@
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>14</maven.compiler.source>
-        <maven.compiler.target>14</maven.compiler.target>
+        <maven.compiler.source>18</maven.compiler.source>
+        <maven.compiler.target>18</maven.compiler.target>
 
         <spring-data-release-train>Neumann-RELEASE</spring-data-release-train>
-        <spring-version>5.2.6.RELEASE</spring-version>
-        <spring-boot-version>2.3.0.RELEASE</spring-boot-version>
-        <spring-mongo-version>3.0.0.RELEASE</spring-mongo-version>
+        <spring-version>5.3.20</spring-version>
+        <spring-boot-version>2.6.7</spring-boot-version>
+        <spring-mongo-version>3.4.0</spring-mongo-version>
         <spring-hateoas-version>1.1.0.RELEASE</spring-hateoas-version>
-        <kotlin.version>1.6.0</kotlin.version>
-        <kotlin.maven.plugin.version>1.6.0-RC</kotlin.maven.plugin.version>
+        <kotlin.version>1.6.21</kotlin.version>
+        <kotlin.maven.plugin.version>1.6.0</kotlin.maven.plugin.version>
     </properties>
 
     <dependencyManagement>
@@ -100,6 +108,7 @@
         <dependency>
             <groupId>org.springframework.data</groupId>
             <artifactId>spring-data-jpa</artifactId>
+            <version>2.7.0</version>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
